requirements.txt
playwright
playwright-stealth
pyyaml
fake-useragent
